+++ config/piano.ts
@@ -1,9 +1,14 @@
 import { Range } from "@utils/typings/Visualizer";
+import { Theme } from "@utils/typings/Theme";
 
 export const TRACK_PLAYING_SPEED = 100;
 export const HORIZONTAL_GAP_BETWEEN_NOTES = 2;
 export const MS_PER_SECOND = 10;
 export const PIANO_HEIGHT = 210;
+export const DEFAULT_THEME: Theme = {
+  naturalColor: "#8ED1FC",
+  accidentalColor: "#FCB900"
+};
 export const RADIUS = 3;
 export const GLOBAL_HEADER_HEIGHT = 50;
 export const MINIMUM_KEYS_IN_READ_MODE = 40;
