+++ utils/typings/Theme.d.ts
@@ -1,0 +1,4 @@
+export interface Theme {
+  naturalColor: string;
+  accidentalColor: string;
+}
